NF - added set_[qf]form methods to Nifti classes

+++ nibabel/spatialimages.py
@@ -302,7 +302,7 @@ class SpatialImage(object):
             # this abstract level, but so far I think all image formats we know
             # do need 4,4.
             # Copy affine to  isolate from environment
-            affine = np.array(affine, copy=True)
+            affine = np.array(affine, copy=True, dtype=np.float)
             if not affine.shape == (4,4):
                 raise ValueError('Affine should be shape 4,4')
         self._affine = affine
